Updates based on Miao and Craig's comments.
